requirements.txt
flask
line-bot-sdk
openai
requests
python-dotenv